refactor: move embedding infrastructure from memory/ to models/embedding/
Move embedder.ts, provider-embedder.ts, and factory.ts out of
memory/embedding/ into models/embedding/ since embeddings are a
general-purpose capability, not memory-specific.

Rename MemoryEmbedder interface to Embedder.
